Refactor graph property calculation to master/worker pattern
